chore: improve cloud structure
