+++ CONTRIBUTING.md
@@ -19,7 +19,7 @@ We recommend developing with the excellent cloud-based development environment p
 
 ### Docker
 
-You can build the Docker image using the `Dockerfile.dev` configuration. If you don't have an existing PostgreSQL service on your machine, you can use our `docker-compose` setup to also start a PostgreSQL database and Redis. Just run `docker-compose -f docker-compose.dev.yml up`.
+You can build the Docker image using the `Dockerfile.dev` configuration. If you don't want to run Postgres or Redis on your machine, you can use our `docke-compose` file to start them via Docker. Just run `docker-compose -f docker-compose.dev.yml up`.
 
 Don't forget to create an `.env` file using `.env.example` as a template.
 
@@ -29,7 +29,7 @@ If you're not using `docker-compose`, make sure to create the database with `doc
 
 Dependencies:
 - ruby ~3.0.0
-- rails ~7.0.0
+- rails ~6.1.3
 - postgres ^13.2
 - redis ^5.07
 
@@ -49,4 +49,4 @@ To start the server, run `.foreman start -f Procfile.dev`. This starts the Rails
 ## Mapzy Cloud
 There are a few places in the codebase where you will encounter logic that is only needed for Mapzy Cloud, the hosted solution by us. These parts of the code are not needed if you are self-hosting, and probably also not relevant for you as a contributor. For the most part this concerns code that relates to payment and subscriptions.
 
-Currently, we use the environment variable `MAPZY_CLOUD` to determine if the instance is being self-hosted or not. Later on, we might improve this structure by factoring out the main code base to a Rails Engine.
+Currently, we use the environment variable `MAPZY_CLOUD` to determine if the instance is being self-hosted or not.
