+++ app/controllers/dashboard/accounts_controller.rb
@@ -3,6 +3,7 @@
 module Dashboard
   class AccountsController < DashboardController
     include Trackable
+    include Cloud::Stripeable
 
     after_action -> { track_event("Viewed Account Settings") }, only: %i[settings]
     after_action -> { track_event("Viewed Embed Code") }, only: %i[embed]
@@ -27,21 +28,6 @@ module Dashboard
 
     def account
       @account ||= current_user.account
-    end
-
-    def stripe_product
-      @stripe_product ||= stripe_subscription.items&.data&.first&.price&.product
-    end
-
-    def stripe_subscription
-      @stripe_subscription ||= stripe_customer.subscriptions&.data&.first
-    end
-
-    def stripe_customer
-      @stripe_customer ||= Stripe::Customer.retrieve(
-        id: current_user.account.stripe_customer_id,
-        expand: ["subscriptions"]
-      )
     end
   end
 end
