+++ app/controllers/users/registrations_controller.rb
@@ -3,6 +3,7 @@
 module Users
   class RegistrationsController < Devise::RegistrationsController
     include Trackable
+    include Cloud::RegistrationWorkable
 
     before_action :configure_sign_up_params, only: [:create]
     before_action :configure_account_update_params, only: [:update]
@@ -16,8 +17,8 @@ module Users
       super do |resource|
         if resource.persisted?
           resource.create_account
-          resource.cloud_emails.setup_email_workers
-          resource.cloud_emails.send_welcome_email
+          send_welcome_email(resource.email)
+          setup_registration_workers(resource.id)
         end
 
         flash.now[:alert] = resource.errors.full_messages.join(", ") if resource.errors.present?
