+++ app/helpers/application_helper.rb
@@ -1,4 +1,7 @@
 # frozen_string_literal: true
 
 module ApplicationHelper
+  def allow_registration?
+    ENV["ALLOW_REGISTRATION"] == "true"
+  end
 end
