+++ app/helpers/cloud_helper.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module CloudHelper
+  def mapzy_cloud?
+    ENV["MAPZY_CLOUD"] == "true"
+  end
+end
